img2oss/requirements.txt
Flask
oss2
requests
Pillow